refactor: make array diff mutations to functional
